requirements.txt
Pillow==9.0.1
tqdm==4.65.0